requirements.txt
annotated-types==0.7.0
appnope==0.1.4
astroid==3.3.10
asttokens==3.0.0
beautifulsoup4==4.13.5
boto3==1.40.35
botocore==1.40.35
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==2.1.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
executing==2.2.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.14
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pip==25.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pylint==3.3.7
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2025.2
pyzmq==27.0.0
requests==2.28.1
s3transfer==0.14.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
